pyproject.toml
[tool.poetry]
name = "de-zoomcamp-project"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "de_zoomcamp_project"}]

[tool.poetry.dependencies]
python = "^3.9"
dbt-core = "^1.4.5"
dbt-bigquery = "^1.4.3"
pyspark = "^3.3.2"
pytest = "^7.2.2"
black = "^23.3.0"
python-dotenv = "^1.0.0"
prefect = "^2.9.0"
gcsfs = "^2023.1.0"
google-cloud-dataproc = "^5.4.1"
prefect-shell = "^0.1.5"
google-cloud-bigquery-datatransfer = "^3.11.1"
pylint = "^2.17.4"
isort = "^5.12.0"
pandas = "^2.0.3"
google-cloud = "^0.34.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
